requirements.txt
click
requests
topology-yaml==0.1.1
pytest
tox
